output/prototypes/mem0_memory_layer/requirements.txt
mem0ai>=0.1.0
openai>=1.0.0
